requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
google-cloud-storage==2.10.0
google-cloud-documentai==3.7.0
google-genai==0.6.1
google-cloud-texttospeech==2.11.0
google-cloud-speech==2.34.0
pydub==0.25.1
aiofiles==23.1.0
